requirements.txt
chumpy==0.70
datasets==2.15.0
einops==0.7.0
face_alignment==1.3.5
imageio==2.19.3
imageio-ffmpeg
ipdb==0.13.13
kornia==0.6.11
loguru==0.7.2
lpips==0.1.4
matplotlib==3.5.2
numpy==1.23.1
opencv_python==4.7.0.72
openmesh==1.2.1
Pillow==10.1.0
pytorch_msssim==1.0.0
PyYAML==6.0.1
scikit_image==0.19.2
scikit_learn==1.0.2
scipy==1.9.1
setuptools==63.4.1
torch==2.0.0
torch_scatter==2.1.2
torchvision==0.15.1
tqdm==4.64.1
yacs==0.1.8
